zyq修改新建信件,编辑信件,插入音频接口

+++ src/server/routes/music.js
@@ -9,7 +9,6 @@ const { getToken, checkToken } = require('../src/token');
 const getRandom = require('../src/user/verification');
 var multiparty = require('multiparty');
 var fs = require('fs');
-
 /**
  * 插入音频
  * 请求方式：
@@ -38,9 +37,14 @@ router.post('/insertMp3', function(req, res){
                     }else{
                         runSql('select music from pletter where pid=?',[pid],(result1)=>{
                             var mp3 = result1.data[0].music;
-                            runSql('update pletter set music=? where pid=? ',[name,pid],(result2)=>{
-                                fs.unlinkSync(path.join(__dirname,'../public/music/'+mp3));
-                            })
+                            if(mp3==null){
+                                runSql('update pletter set music=? where pid=? ',[name,pid],(result2)=>{
+                                })
+                            }else{
+                                runSql('update pletter set music=? where pid=? ',[name,pid],(result3)=>{
+                                    fs.unlinkSync(path.join(__dirname,'../public/music/'+mp3));
+                                })
+                            }
                             res.json({status: 0, data: [name]});
                         })
                     }
@@ -85,7 +89,6 @@ router.get('/showmusic',function(req,res){
  *      pid：信件id
  *      music：音乐名
  */
-
 router.post('/delMp3',function(req,res,next){
     let token = req.header('token');
     let {pid,music} = req.body;
@@ -93,7 +96,7 @@ router.post('/delMp3',function(req,res,next){
         if(result.status !==0){
             res.json(result)
         }else{
-            runSql('update pletter set music=? where pid=?',['',pid],(result2)=>{
+            runSql('update pletter set music=? where pid=?',[null,pid],(result2)=>{
                 fs.unlinkSync(path.join(__dirname,'../public/music/'+music));
                 res.json(result2);
             })
